REF: Move add_jitter and add_noise to sim.distortion
These distortion functions should be in the simulation module
with the other distortions. Also, add_noise is deprecated and
marked for future removal because it is redundant with
add_gaussian.
