+++ source/tomopy/sim/distortion.py
@@ -74,6 +74,8 @@ __all__ = [
     'add_focal_spot_blur',
     'add_rings',
     'add_zingers',
+    'add_jitter',
+    'add_noise',
 ]
 
 
@@ -97,9 +99,78 @@ def add_gaussian(tomo, mean=0, std=None):
     tomo = dtype.as_ndarray(tomo)
     if std is None:
         std = tomo.max() * 0.05
-    dx, dy, dz = tomo.shape
-    tomo += std * np.random.randn(dx, dy, dz) + mean
-    return tomo
+    noise = np.random.normal(mean, std, size=tomo.shape)
+    return tomo + noise
+
+
+def add_jitter(prj, low=0, high=1):
+    """
+    Simulates jitter in projection images. The jitter
+    is simulated by drawing random samples from a uniform
+    distribution over the half-open interval [low, high).
+
+    Parameters
+    ----------
+    prj : ndarray
+        3D stack of projection images. The first dimension
+        is projection axis, second and third dimensions are
+        the x- and y-axes of the projection image, respectively.
+    low : float, optional
+        Lower boundary of the output interval. All values
+        generated will be greater than or equal to low. The
+        default value is 0.
+    high : float
+        Upper boundary of the output interval. All values
+        generated will be less than high. The default value
+        is 1.0.
+
+    Returns
+    -------
+    ndarray
+        3D stack of projection images with jitter.
+    """
+    from skimage import transform as tf
+
+    # Needs scaling for skimage float operations.
+    prj, scl = scale(prj)
+
+    # Random jitter parameters are drawn from uniform distribution.
+    jitter = np.random.uniform(low, high, size=(prj.shape[0], 2))
+
+    for m in range(prj.shape[0]):
+        tform = tf.SimilarityTransform(translation=jitter[m])
+        prj[m] = tf.warp(prj[m], tform, order=0)
+
+    # Re-scale back to original values.
+    prj *= scl
+    return prj, jitter[:, 0], jitter[:, 1]
+
+
+def add_noise(prj, ratio=0.05):
+    """
+    Adds Gaussian noise with zero mean and a given standard
+    deviation as a ratio of the maximum value in data.
+
+    Parameters
+    ----------
+    prj : ndarray
+        3D stack of projection images. The first dimension
+        is projection axis, second and third dimensions are
+        the x- and y-axes of the projection image, respectively.
+    ratio : float, optional
+        Ratio of the standard deviation of the Gaussian noise
+        distribution to the maximum value in data.
+
+    Returns
+    -------
+    ndarray
+        3D stack of projection images with added Gaussian noise.
+    """
+    warnings.warn("'add_noise' will be removed in a future version."
+                  " Use 'add_gaussian' instead.", FutureWarning)
+    std = prj.max() * ratio
+    noise = np.random.normal(0, std, size=prj.shape)
+    return prj + noise.astype('float32')
 
 
 def add_poisson(tomo):
